jobs/upload_to_gcs/pyproject.toml
[project]
name = "upload-to-gcs"
version = "0.1.0"
description = ""
authors = [
    {name = "Lasnami Abdelkarim",email = "81482604+LasnamiWalid16@users.noreply.github.com"}
]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "google-cloud-storage (>=3.1.0,<4.0.0)",
    "python-dotenv (>=1.1.0,<2.0.0)",
    "pre-commit (>=4.2.0,<5.0.0)",
    "dotenv (>=0.9.9,<0.10.0)"
]

[tool.poetry]
packages = [{include = "upload_to_gcs", from = "src"}]


[tool.poetry.group.dev.dependencies]
pytest = "^8.3.5"
pytest-mock = "^3.14.0"
python-dotenv = "^1.1.0"
pre-commit = "^4.2.0"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
